pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rloo-project"
version = "0.1.0"
description = "Minimal RLOO fine-tuning setup for code completion"
readme = "README.md"
authors = [{ name = "Your Name" }]
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.26",
    "datasets>=2.14",
    "evaluate>=0.4",
    "numpy>=1.23",
    "peft>=0.7",
    "torch",
    "transformers>=4.38",
    "trl>=0.9",
    "yamlargparse>=1.26",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "black>=24.3",
    "ruff>=0.5",
    "pytest>=7.4"
]

[tool.setuptools]
packages = ["rloo_project"]
package-dir = {"" = "src"}
